📝 (Improve readmes) Add descriptive readme for each section and whole repo
Accessing this repository was not a very pleasant experience since there was basically nothing to help navigation, and no description of the repo or any of its sections.

This commit adds short but descriptive READMEs to each section, and also adds a main project README file with more information about the project.