✨ (Missing sections) Add short type hint guide for section 8
Section 8 was missing in the repo, but it didn't make sense to add code files as in the videos it uses existing code to explain some concepts.

Instead I added some explanations with code blocks, to act as reminders (a cheatsheet, if you will) of what sort of thing you can do with type hints.